changed formatting to ratingView fixed a bug where I encountered pricing in listingitemview as 120.000000

+++ AirbnbTutorial/Core/Listing/Model/Listing.swift
@@ -16,7 +16,7 @@ struct Listing: Identifiable, Codable, Hashable {
     let noOfBathrooms: Int
     let noOfGuests: Int
     let noOfBeds: Int
-    var pricePerNight: Double
+    var pricePerNight: Int
     let latitude: Double
     let longitude: Double
     var imageURL: [String]
